# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TreeKB API
  version: "2.3.0"
  description: |
    TreeKB is an AI-driven, self-evolving knowledge graph system. It ingests documents,
    extracts knowledge points (KPs), organises them into a hierarchical topic tree, and
    provides intelligent question-answering with citations.

    **Key concepts**
    - **Dataset** -- an isolated knowledge base with its own SQLite database.
    - **Node** -- a topic node in the hierarchical tree.
    - **Chunk / KP** -- an atomic knowledge point stored under a node.
    - **Schema node** -- a curated node that guides ingestion mapping.
    - **Decision** -- a pending human-review item produced by the KP decision engine.

    Most endpoints require the `X-Dataset-ID` header to select the active dataset.
    Endpoints under `/datasets`, `/health`, `/metrics`, and `/settings` do not.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  parameters:
    DatasetIdHeader:
      name: X-Dataset-ID
      in: header
      required: true
      description: UUID of the target dataset. Resolved by middleware; falls back to default dataset when omitted.
      schema:
        type: string
        format: uuid
  schemas:
...
    SimpleError:
      type: object
      properties:
        error:
          type: string

    # ── Dataset ───────────────────────────────────────────────────────────────
...
    IngestionJob:
      type: object
      properties:
        id:
          type: integer
        file_path:
          type: string
        original_name:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed, rate_limited, cancelled]
        attempt_count:
          type: integer
        max_attempts:
          type: integer
        error_message:
          type: string
          nullable: true
        document_id:
          type: integer
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Decision ──────────────────────────────────────────────────────────────
...
paths:
...
  /upload:
    post:
      operationId: uploadFile
      tags: [Ingestion]
      summary: Upload a single file
      description: |
        Uploads and ingests a single document. By default the file is queued for asynchronous background processing
        (returns 202). Set `sync=true` in the body to process synchronously (returns 200 with result).
        Supported file types include PDF, DOCX, TXT, MD, HTML, CSV, XLSX.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
                targetNodeId:
                  type: string
                  description: Force-assign to a specific tree node
                targetSchemaNodeId:
                  type: string
                  description: Assign to a schema node
                useLLM:
                  type: boolean
                  default: true
                  description: Whether to use LLM for extraction
                sync:
                  type: boolean
                  default: false
                  description: Process synchronously (blocks until complete)
      responses:
        "200":
          description: Synchronous processing result (when sync=true)
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: integer
                  chunks_created:
                    type: integer
                  nodes_created:
                    type: integer
        "202":
          description: File queued for background processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:
                    type: boolean
                  message:
                    type: string
                  job:
                    $ref: "#/components/schemas/IngestionJob"
        "400":
          description: No file uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /upload/batch:
    post:
      operationId: uploadBatch
      tags: [Ingestion]
      summary: Upload multiple files
      description: |
        Uploads and ingests multiple documents at once. By default files are queued for asynchronous processing
        (returns 202). Set `sync=true` to process all files synchronously.
        Maximum file count controlled by INGEST_MAX_BATCH_FILES env var (default 100).
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Array of files to upload
                targetNodeId:
                  type: string
                targetSchemaNodeId:
                  type: string
                useLLM:
                  type: boolean
                  default: true
                sync:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Synchronous processing result (when sync=true)
          content:
            application/json:
              schema:
                type: object
        "202":
          description: Files queued for background processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:
                    type: boolean
                  message:
                    type: string
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/IngestionJob"
                  count:
                    type: integer
        "400":
          description: No files uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /ingest/jobs:
    get:
      operationId: listIngestionJobs
      tags: [Ingestion]
      summary: List ingestion jobs
      description: Returns ingestion jobs, optionally filtered by status, with pagination.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: status
          in: query
          description: Filter by job status
          schema:
            type: string
            enum: [queued, processing, completed, failed, rate_limited, cancelled]
        - name: limit
          in: query
          description: Maximum results (1-200)
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/IngestionJob"
                  count:
                    type: integer
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /ingest/jobs/active:
    get:
      operationId: listActiveIngestionJobs
      tags: [Ingestion]
      summary: List active ingestion jobs
      description: Returns only active (queued, processing, rate_limited) ingestion jobs.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Active job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/IngestionJob"
                  count:
                    type: integer
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /ingest/jobs/cancel-all:
    post:
      operationId: cancelAllIngestionJobs
      tags: [Ingestion]
      summary: Cancel all pending jobs
      description: Cancels all queued and rate-limited ingestion jobs.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Jobs cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cancelled:
                    type: integer
  /ingest/jobs/retry-all:
    post:
      operationId: retryAllIngestionJobs
      tags: [Ingestion]
      summary: Retry all failed jobs
      description: Resets all failed and rate-limited jobs to queued status for reprocessing.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Jobs retried
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  retried:
                    type: integer
  /ingest/jobs/completed:
    delete:
      operationId: purgeCompletedJobs
      tags: [Ingestion]
      summary: Purge completed jobs
      description: Deletes all completed job records from the ingestion_jobs table.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Completed jobs purged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  deleted:
                    type: integer
  /ingest/jobs/failed:
    delete:
      operationId: purgeFailedJobs
      tags: [Ingestion]
      summary: Purge failed jobs
      description: Deletes all failed job records from the ingestion_jobs table.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Failed jobs purged
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  deleted:
                    type: integer
  /ingest/jobs/{id}:
    get:
      operationId: getIngestionJob
      tags: [Ingestion]
      summary: Get a single ingestion job
      description: Returns details of a specific ingestion job by ID.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /ingest/jobs/{id}/retry:
    post:
      operationId: retryIngestionJob
      tags: [Ingestion]
      summary: Retry a specific job
      description: Resets a failed or rate-limited job to queued status for reprocessing.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job retried
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  job:
                    $ref: "#/components/schemas/IngestionJob"
        "400":
          description: Job cannot be retried
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /ingest/jobs/{id}/cancel:
    post:
      operationId: cancelIngestionJob
      tags: [Ingestion]
      summary: Cancel a specific job
      description: Cancels a queued or rate-limited ingestion job.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  job:
                    $ref: "#/components/schemas/IngestionJob"
        "400":
          description: Job cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
  /ingest/queue/stats:
    get:
      operationId: getIngestionQueueStats
      tags: [Ingestion]
      summary: Get ingestion queue statistics
      description: Returns counts of jobs by status in the ingestion queue.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  queued:
                    type: integer
                  processing:
                    type: integer
                  completed:
                    type: integer
                  failed:
                    type: integer
                  rate_limited:
                    type: integer
                  cancelled:
                    type: integer

  # ── Documents ─────────────────────────────────────────────────────────────
...