# src/routes/ingest.js
import express from "express";
import multer from "multer";
import path from "path";
import fs from "fs";
import { fileURLToPath } from "url";
import { processDocument, processDocumentBatch, isSupportedFileType, getSupportedExtensions } from "../ingest/index.js";
import {
  enqueueIngestionJob,
  enqueueIngestionJobs,
  getIngestionJob,
  listIngestionJobs,
  listActiveIngestionJobs,
  retryIngestionJob,
  cancelIngestionJob,
  cancelAllIngestionJobs,
  retryAllIngestionJobs,
  getIngestionQueueStats
} from "../ingest/jobQueue.js";
import { JobRepo } from "../db/repositories/JobRepo.js";
import { runWithDb } from "../db/activeDb.js";
import { apiLogger } from "../utils/logger.js";
import { ApiError } from "../utils/apiError.js";

const __dirname = path.dirname(fileURLToPath(import.meta.url));

const router = express.Router();

// Configure multer for file uploads
const uploadDir = path.join(__dirname, "../../uploads");
if (!fs.existsSync(uploadDir)) {
  fs.mkdirSync(uploadDir, { recursive: true });
}

const storage = multer.diskStorage({
  destination: (req, file, cb) => cb(null, uploadDir),
  filename: (req, file, cb) => {
    const ext = path.extname(file.originalname).toLowerCase() || "";
    const uniqueName = `${Date.now()}-${Math.random().toString(36).slice(2)}${ext}`;
    cb(null, uniqueName);
  }
});

const upload = multer({
  storage,
  fileFilter: (req, file, cb) => {
    // Multer reads originalname as latin1 per HTTP spec; browsers send UTF-8.
    // Re-decode so Chinese (and other non-ASCII) names display correctly.
    file.originalname = Buffer.from(file.originalname, "latin1").toString("utf8");
    if (isSupportedFileType(file.originalname)) {
      cb(null, true);
    } else {
      cb(new Error(`Unsupported file type. Supported: ${getSupportedExtensions().join(", ")}`));
    }
  },
  limits: { fileSize: (Number(process.env.INGEST_MAX_FILE_MB) || 200) * 1024 * 1024 }
});

// ==================== UPLOAD ====================

// Upload and process a single file
router.post("/upload", upload.single("file"), async (req, res) => {
  // Re-establish dataset context: multer's async processing breaks AsyncLocalStorage
  await runWithDb(req.datasetConn, async () => {
    try {
      if (!req.file) return res.status(400).json({ error: { code: 'VALIDATION_ERROR', message: "No file uploaded" } });

      const { targetNodeId, targetSchemaNodeId, useLLM = true, sync = false } = req.body || {};
      const processOptions = {
        targetNodeId,
        targetSchemaNodeId: targetSchemaNodeId || null,
        useLLM: useLLM === "true" || useLLM === true,
        originalName: req.file.originalname
      };

      if (sync === "true" || sync === true) {
        let result;
        try {
          result = await processDocument(req.file.path, processOptions);
        } catch (procErr) {
          try { fs.unlinkSync(req.file.path); } catch (_) {}
          throw procErr;
        }
        try { fs.unlinkSync(req.file.path); } catch (_) {}
        return res.json(result);
      }

      const job = enqueueIngestionJob(req.file.path, {
        originalName: req.file.originalname,
        fileSize: req.file.size,
        processOptions
      });

      res.status(202).json({
        queued: true,
        message: "File queued for background processing",
        job
      });
    } catch (err) {
      if (req.file?.path) { try { fs.unlinkSync(req.file.path); } catch (_) {} }
      if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
      apiLogger.error("Upload error:", err.message);
      res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
    }
  });
});

// Upload and process multiple files
router.post("/upload/batch", upload.array("files", Number(process.env.INGEST_MAX_BATCH_FILES) || 100), async (req, res) => {
  // Re-establish dataset context: multer's async processing breaks AsyncLocalStorage
  await runWithDb(req.datasetConn, async () => {
    try {
      if (!req.files || req.files.length === 0) {
        return res.status(400).json({ error: { code: 'VALIDATION_ERROR', message: "No files uploaded" } });
      }

      const { targetNodeId, targetSchemaNodeId, useLLM = true, sync = false } = req.body || {};
      const processOptions = {
        targetNodeId,
        targetSchemaNodeId: targetSchemaNodeId || null,
        useLLM: useLLM === "true" || useLLM === true
      };

      if (sync === "true" || sync === true) {
        let result;
        try {
          const filePaths = req.files.map(f => f.path);
          result = await processDocumentBatch(filePaths, processOptions);
        } catch (procErr) {
          // Clean up uploaded files on processing error
          for (const f of req.files) { try { fs.unlinkSync(f.path); } catch (_) {} }
          throw procErr;
        }
        // Clean up uploaded files after successful sync processing
        for (const f of req.files) { try { fs.unlinkSync(f.path); } catch (_) {} }
        return res.json(result);
      }

      const jobs = enqueueIngestionJobs(req.files, { processOptions });
      res.status(202).json({
        queued: true,
        message: `${jobs.length} files queued for background processing`,
        jobs,
        count: jobs.length
      });
    } catch (err) {
      // Clean up uploaded files on error
      if (Array.isArray(req.files)) {
        for (const f of req.files) { try { fs.unlinkSync(f.path); } catch (_) {} }
      }
      if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
      apiLogger.error("Batch upload error:", err.message);
      res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
    }
  });
});

// ==================== INGESTION JOBS ====================

router.get("/ingest/jobs/active", (req, res) => {
  try {
    const jobs = listActiveIngestionJobs();
    res.json({ jobs, count: jobs.length });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("List active jobs error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.post("/ingest/jobs/cancel-all", (req, res) => {
  try {
    const cancelled = cancelAllIngestionJobs();
    res.json({ success: true, cancelled });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Cancel all jobs error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.post("/ingest/jobs/retry-all", (req, res) => {
  try {
    const retried = retryAllIngestionJobs();
    res.json({ success: true, retried });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Retry all jobs error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.delete("/ingest/jobs/completed", (req, res) => {
  try {
    const deleted = JobRepo.purgeCompleted();
    res.json({ ok: true, deleted });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Purge completed jobs error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.delete("/ingest/jobs/failed", (req, res) => {
  try {
    const deleted = JobRepo.purgeFailed();
    res.json({ ok: true, deleted });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Purge failed jobs error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.get("/ingest/jobs", (req, res) => {
  try {
    const { status } = req.query;
    const limit = Math.min(200, Math.max(1, parseInt(req.query.limit, 10) || 50));
    const offset = Math.max(0, parseInt(req.query.offset, 10) || 0);
    const jobs = listIngestionJobs({ status, limit, offset });
    res.json({ jobs, count: jobs.length });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("List ingestion jobs error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.get("/ingest/jobs/:id", (req, res) => {
  try {
    const job = getIngestionJob(parseInt(req.params.id));
    if (!job) return res.status(404).json({ error: { code: 'NOT_FOUND', message: "Ingestion job not found" } });
    res.json(job);
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Get ingestion job error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

router.post("/ingest/jobs/:id/retry", (req, res) => {
  try {
    const job = retryIngestionJob(parseInt(req.params.id));
    res.json({ success: true, job });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Retry ingestion job error:", err.message);
    res.status(400).json({ error: { code: 'VALIDATION_ERROR', message: err.message } });
  }
});

router.post("/ingest/jobs/:id/cancel", (req, res) => {
  try {
    const job = cancelIngestionJob(parseInt(req.params.id));
    res.json({ success: true, job });
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Cancel ingestion job error:", err.message);
    res.status(400).json({ error: { code: 'VALIDATION_ERROR', message: err.message } });
  }
});

router.get("/ingest/queue/stats", (req, res) => {
  try {
    res.json(getIngestionQueueStats());
  } catch (err) {
    if (err instanceof ApiError) return res.status(err.status).json(err.toJSON());
    apiLogger.error("Get ingestion queue stats error:", err.message);
    res.status(500).json({ error: { code: 'INTERNAL_ERROR', message: err.message } });
  }
});

export default router;